pageturners-backend/requirements.txt
Flask>=3.0.0
Flask-CORS>=5.0.0
python-dotenv>=1.0.1
pymongo>=4.6.0
bcrypt>=4.1.0
requests>=2.31.0
Flask-JWT-Extended>=4.6.0
pytest>=8.0.0
mongomock>=4.1.0
